# docs/api-contract.yaml
openapi: '3.0.3'
info:
  title: TrustLedger API
  description: |
    MSMED Act compliance platform for Indian MSMEs.
    All invoice endpoints require Firebase Authentication.
  version: 1.0.0
  contact:
    name: TrustLedger Team
  license:
    name: Proprietary
servers:
  - url: /api
    description: API base path (set via VITE_API_URL / MONGO_URL)
paths:
...
  /users/check-lockout:
    post:
      summary: Check if an account is locked or CAPTCHA is required
      operationId: checkLockout
      tags: [Users, Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: msme@example.com
      responses:
        '200':
          description: Lockout and CAPTCHA status
          content:
            application/json:
              schema:
                type: object
                properties:
                  locked:
                    type: boolean
                  remainingMinutes:
                    type: integer
                  captchaRequired:
                    type: boolean
  /users/record-failed-attempt:
    post:
      summary: Record a failed login attempt (triggers lockout at threshold)
      operationId: recordFailedAttempt
      tags: [Users, Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Attempt recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /users/reset-attempts:
    post:
      summary: Reset failed-attempt counters and track new device login
      operationId: resetAttempts
      tags: [Users, Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Counters reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /users/verify-captcha:
    post:
      summary: Verify a reCAPTCHA token
      operationId: verifyCaptcha
      tags: [Users, Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: reCAPTCHA response token
      responses:
        '200':
          description: CAPTCHA verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '400':
          description: Verification failed
  /users/me:
    get:
      summary: Get the authenticated user's profile
      operationId: getProfile
      tags: [Users, Profile]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found

    post:
      summary: Create a profile for the authenticated user (idempotent)
      operationId: createProfile
      tags: [Users, Profile]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 256
                email:
                  type: string
                  format: email
                authProvider:
                  type: string
                  enum: [email, google]
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '200':
          description: Profile already exists (idempotent)

    patch:
      summary: Update the authenticated user's profile (whitelisted fields only)
      operationId: updateProfile
      tags: [Users, Profile]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 256
                companyName:
                  type: string
                udyamNumber:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                profileComplete:
                  type: boolean
                emailReminders:
                  type: boolean
                dataSharing:
                  type: boolean
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/pin/status:
    get:
      summary: Check whether the user has a PIN set
      operationId: getPinStatus
      tags: [Users, PIN]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: PIN status
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasPin:
                    type: boolean
  /users/pin/set:
    post:
      summary: Set or overwrite the app PIN
      operationId: setPin
      tags: [Users, PIN]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin]
              properties:
                pin:
                  type: string
                  pattern: '^\d{4}$'
                  description: Exactly 4 digits
      responses:
        '200':
          description: PIN set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '400':
          description: Invalid PIN format
  /users/pin/verify:
    post:
      summary: Verify the app PIN (rate-limited, 5 attempts → 60s lockout)
      operationId: verifyPin
      tags: [Users, PIN]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin]
              properties:
                pin:
                  type: string
                  pattern: '^\d{4}$'
      responses:
        '200':
          description: PIN correct
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '401':
          description: Incorrect PIN
        '429':
          description: Too many attempts (locked)
  /users/pin/change:
    post:
      summary: Change the PIN (requires current PIN)
      operationId: changePin
      tags: [Users, PIN]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPin, newPin]
              properties:
                currentPin:
                  type: string
                  pattern: '^\d{4}$'
                newPin:
                  type: string
                  pattern: '^\d{4}$'
      responses:
        '200':
          description: PIN changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '401':
          description: Incorrect current PIN
  /users/pin/reset:
    post:
      summary: Force-reset the PIN (requires CAPTCHA if IP threshold is met)
      operationId: resetPin
      tags: [Users, PIN]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin]
              properties:
                pin:
                  type: string
                  pattern: '^\d{4}$'
                captchaToken:
                  type: string
                  description: reCAPTCHA token required if suspicious IP
      responses:
        '200':
          description: PIN reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '400':
          description: CAPTCHA required or invalid
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Auth ID token
...
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        firebaseUid:
          type: string
        name:
          type: string
        companyName:
          type: string
        udyamNumber:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        authProvider:
          type: string
          enum: [email, google]
        profileComplete:
          type: boolean
        pinHash:
          type: string
        pinAttempts:
          type: integer
        pinLockedUntil:
          type: string
          nullable: true
          format: date-time
        passwordAttempts:
          type: integer
        passwordLockedUntil:
          type: string
          nullable: true
          format: date-time
        emailReminders:
          type: boolean
        dataSharing:
          type: boolean
        knownDevices:
          type: array
          items:
            type: string
        lastLoginAt:
          type: string
          nullable: true
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...